src/requirements.txt
# Intentionally outdated versions for security scanning demonstration
# DO NOT USE IN PRODUCTION
flask==2.0.1
requests==2.20.0
pyyaml==5.3.1
sqlalchemy==1.3.0
